+++ factory/__init__.py
@@ -43,6 +43,7 @@ from .base import (
 # Backward compatibility; this should be removed soon.
 from .mogo import MogoFactory
 from .django import DjangoModelFactory
+from .faker import Faker
 
 from .declarations import (
     LazyAttribute,