+++ tests/__init__.py
@@ -7,6 +7,7 @@ from .test_django import *
 from .test_base import *
 from .test_containers import *
 from .test_declarations import *
+from .test_faker import *
 from .test_fuzzy import *
 from .test_helpers import *
 from .test_using import *